requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.5
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Flask-CORS==4.0.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
supabase==2.3.4
gunicorn==21.2.0
eventlet==0.33.3
qrcode==7.4.2
Pillow==10.1.0
Werkzeug==3.0.1
eventlet==0.35.2
